requirements.txt
streamlit>=1.39
python-dotenv>=1.0
pydantic>=2.7
groq>=0.18
reportlab>=4.2
